refactor(project-files): extract mutation owner
